separate taxonomy utilities from smasher (move changes from registry branch)
