+++ t/aster.py
@@ -1,10 +1,11 @@
 # Jython script to build the "model village" taxonomy.
 
-from org.opentreeoflife.smasher import Taxonomy
+from org.opentreeoflife.taxa import Taxonomy
+from org.opentreeoflife.smasher import UnionTaxonomy
 from claim import Has_child
 
 # Create model taxonomy
-tax = Taxonomy.newTaxonomy()
+tax = UnionTaxonomy()
 
 # Establish homonym-resolution skeleton (not really used here)
 # skel = Taxonomy.getTaxonomy('tax/skel/', 'skel')
